improve single responsibility of ZipkinSpanExporter by factoring out OtelToZipkinSpanTransformer.
