+++ exporters/zipkin/src/main/java/io/opentelemetry/exporter/zipkin/ZipkinSpanExporter.java
@@ -5,36 +5,20 @@
 
 package io.opentelemetry.exporter.zipkin;
 
-import static io.opentelemetry.api.common.AttributeKey.stringKey;
-import static java.util.concurrent.TimeUnit.NANOSECONDS;
-
-import io.opentelemetry.api.common.AttributeKey;
-import io.opentelemetry.api.common.AttributeType;
-import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.metrics.MeterProvider;
-import io.opentelemetry.api.trace.StatusCode;
 import io.opentelemetry.exporter.internal.ExporterMetrics;
 import io.opentelemetry.sdk.common.CompletableResultCode;
-import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
 import io.opentelemetry.sdk.internal.ThrottlingLogger;
-import io.opentelemetry.sdk.resources.Resource;
-import io.opentelemetry.sdk.trace.data.EventData;
 import io.opentelemetry.sdk.trace.data.SpanData;
-import io.opentelemetry.sdk.trace.data.StatusData;
 import io.opentelemetry.sdk.trace.export.SpanExporter;
-import io.opentelemetry.semconv.resource.attributes.ResourceAttributes;
 import java.io.IOException;
-import java.net.InetAddress;
-import java.net.NetworkInterface;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Enumeration;
 import java.util.List;
+import java.util.function.Function;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import javax.annotation.Nullable;
 import zipkin2.Callback;
-import zipkin2.Endpoint;
 import zipkin2.Span;
 import zipkin2.codec.BytesEncoder;
 import zipkin2.codec.Encoding;
@@ -51,185 +35,26 @@ public final class ZipkinSpanExporter implements SpanExporter {
 
   public static final String DEFAULT_ENDPOINT = "http://localhost:9411/api/v2/spans";
 
-  static final String OTEL_DROPPED_ATTRIBUTES_COUNT = "otel.dropped_attributes_count";
-  static final String OTEL_DROPPED_EVENTS_COUNT = "otel.dropped_events_count";
-  static final String OTEL_STATUS_CODE = "otel.status_code";
-  static final AttributeKey<String> STATUS_ERROR = stringKey("error");
-
-  static final String KEY_INSTRUMENTATION_SCOPE_NAME = "otel.scope.name";
-  static final String KEY_INSTRUMENTATION_SCOPE_VERSION = "otel.scope.version";
-  static final String KEY_INSTRUMENTATION_LIBRARY_NAME = "otel.library.name";
-  static final String KEY_INSTRUMENTATION_LIBRARY_VERSION = "otel.library.version";
-
   private final ThrottlingLogger logger = new ThrottlingLogger(baseLogger);
-
   private final BytesEncoder<Span> encoder;
   private final Sender sender;
   private final ExporterMetrics exporterMetrics;
-  @Nullable private final InetAddress localAddress;
+
+  private final Function<SpanData, Span> otelToZipkin;
 
   ZipkinSpanExporter(BytesEncoder<Span> encoder, Sender sender, MeterProvider meterProvider) {
+    this(encoder, sender, new OtelToZipkinSpanTransformer());
+  }
+
+  ZipkinSpanExporter(
+      BytesEncoder<Span> encoder, Sender sender, Function<SpanData, Span> otelToZipkin) {
     this.encoder = encoder;
     this.sender = sender;
     this.exporterMetrics =
         sender.encoding() == Encoding.JSON
             ? ExporterMetrics.createHttpJson("zipkin", "span", meterProvider)
             : ExporterMetrics.createHttpProtobuf("zipkin", "span", meterProvider);
-    localAddress = produceLocalIp();
-  }
-
-  /** Logic borrowed from brave.internal.Platform.produceLocalEndpoint */
-  @Nullable
-  static InetAddress produceLocalIp() {
-    try {
-      Enumeration<NetworkInterface> nics = NetworkInterface.getNetworkInterfaces();
-      while (nics.hasMoreElements()) {
-        NetworkInterface nic = nics.nextElement();
-        Enumeration<InetAddress> addresses = nic.getInetAddresses();
-        while (addresses.hasMoreElements()) {
-          InetAddress address = addresses.nextElement();
-          if (address.isSiteLocalAddress()) {
-            return address;
-          }
-        }
-      }
-    } catch (Exception e) {
-      // don't crash the caller if there was a problem reading nics
-      baseLogger.log(Level.FINE, "error reading nics", e);
-    }
-    return null;
-  }
-
-  // VisibleForTesting
-  Span generateSpan(SpanData spanData) {
-    Endpoint endpoint = getEndpoint(spanData);
-
-    long startTimestamp = toEpochMicros(spanData.getStartEpochNanos());
-    long endTimestamp = toEpochMicros(spanData.getEndEpochNanos());
-
-    Span.Builder spanBuilder =
-        Span.newBuilder()
-            .traceId(spanData.getTraceId())
-            .id(spanData.getSpanId())
-            .kind(toSpanKind(spanData))
-            .name(spanData.getName())
-            .timestamp(toEpochMicros(spanData.getStartEpochNanos()))
-            .duration(Math.max(1, endTimestamp - startTimestamp))
-            .localEndpoint(endpoint);
-
-    if (spanData.getParentSpanContext().isValid()) {
-      spanBuilder.parentId(spanData.getParentSpanId());
-    }
-
-    Attributes spanAttributes = spanData.getAttributes();
-    spanAttributes.forEach(
-        (key, value) -> spanBuilder.putTag(key.getKey(), valueToString(key, value)));
-    int droppedAttributes = spanData.getTotalAttributeCount() - spanAttributes.size();
-    if (droppedAttributes > 0) {
-      spanBuilder.putTag(OTEL_DROPPED_ATTRIBUTES_COUNT, String.valueOf(droppedAttributes));
-    }
-
-    StatusData status = spanData.getStatus();
-
-    // include status code & error
-    if (status.getStatusCode() != StatusCode.UNSET) {
-      spanBuilder.putTag(OTEL_STATUS_CODE, status.getStatusCode().toString());
-
-      // add the error tag, if it isn't already in the source span
-      if (status.getStatusCode() == StatusCode.ERROR && spanAttributes.get(STATUS_ERROR) == null) {
-        spanBuilder.putTag(STATUS_ERROR.getKey(), status.getDescription());
-      }
-    }
-
-    InstrumentationScopeInfo instrumentationScopeInfo = spanData.getInstrumentationScopeInfo();
-
-    if (!instrumentationScopeInfo.getName().isEmpty()) {
-      spanBuilder.putTag(KEY_INSTRUMENTATION_SCOPE_NAME, instrumentationScopeInfo.getName());
-      // include instrumentation library name for backwards compatibility
-      spanBuilder.putTag(KEY_INSTRUMENTATION_LIBRARY_NAME, instrumentationScopeInfo.getName());
-    }
-    if (instrumentationScopeInfo.getVersion() != null) {
-      spanBuilder.putTag(KEY_INSTRUMENTATION_SCOPE_VERSION, instrumentationScopeInfo.getVersion());
-      // include instrumentation library name for backwards compatibility
-      spanBuilder.putTag(
-          KEY_INSTRUMENTATION_LIBRARY_VERSION, instrumentationScopeInfo.getVersion());
-    }
-
-    for (EventData annotation : spanData.getEvents()) {
-      spanBuilder.addAnnotation(toEpochMicros(annotation.getEpochNanos()), annotation.getName());
-    }
-    int droppedEvents = spanData.getTotalRecordedEvents() - spanData.getEvents().size();
-    if (droppedEvents > 0) {
-      spanBuilder.putTag(OTEL_DROPPED_EVENTS_COUNT, String.valueOf(droppedEvents));
-    }
-
-    return spanBuilder.build();
-  }
-
-  private Endpoint getEndpoint(SpanData spanData) {
-    Attributes resourceAttributes = spanData.getResource().getAttributes();
-
-    Endpoint.Builder endpoint = Endpoint.newBuilder().ip(localAddress);
-
-    // use the service.name from the Resource, if it's been set
-    String serviceNameValue = resourceAttributes.get(ResourceAttributes.SERVICE_NAME);
-    if (serviceNameValue == null) {
-      serviceNameValue = Resource.getDefault().getAttribute(ResourceAttributes.SERVICE_NAME);
-    }
-    // in practice should never be null unless the default Resource spec is changed
-    if (serviceNameValue != null) {
-      endpoint.serviceName(serviceNameValue);
-    }
-    return endpoint.build();
-  }
-
-  @Nullable
-  private static Span.Kind toSpanKind(SpanData spanData) {
-    switch (spanData.getKind()) {
-      case SERVER:
-        return Span.Kind.SERVER;
-      case CLIENT:
-        return Span.Kind.CLIENT;
-      case PRODUCER:
-        return Span.Kind.PRODUCER;
-      case CONSUMER:
-        return Span.Kind.CONSUMER;
-      case INTERNAL:
-        return null;
-    }
-    return null;
-  }
-
-  private static long toEpochMicros(long epochNanos) {
-    return NANOSECONDS.toMicros(epochNanos);
-  }
-
-  private static String valueToString(AttributeKey<?> key, Object attributeValue) {
-    AttributeType type = key.getType();
-    switch (type) {
-      case STRING:
-      case BOOLEAN:
-      case LONG:
-      case DOUBLE:
-        return String.valueOf(attributeValue);
-      case STRING_ARRAY:
-      case BOOLEAN_ARRAY:
-      case LONG_ARRAY:
-      case DOUBLE_ARRAY:
-        return commaSeparated((List<?>) attributeValue);
-    }
-    throw new IllegalStateException("Unknown attribute type: " + type);
-  }
-
-  private static String commaSeparated(List<?> values) {
-    StringBuilder builder = new StringBuilder();
-    for (Object value : values) {
-      if (builder.length() != 0) {
-        builder.append(',');
-      }
-      builder.append(value);
-    }
-    return builder.toString();
+    this.otelToZipkin = otelToZipkin;
   }
 
   @Override
@@ -239,7 +64,8 @@ public final class ZipkinSpanExporter implements SpanExporter {
 
     List<byte[]> encodedSpans = new ArrayList<>(numItems);
     for (SpanData spanData : spanDataList) {
-      encodedSpans.add(encoder.encode(generateSpan(spanData)));
+      Span zipkinSpan = otelToZipkin.apply(spanData);
+      encodedSpans.add(encoder.encode(zipkinSpan));
     }
 
     CompletableResultCode result = new CompletableResultCode();
@@ -286,11 +112,5 @@ public final class ZipkinSpanExporter implements SpanExporter {
    */
   public static ZipkinSpanExporterBuilder builder() {
     return new ZipkinSpanExporterBuilder();
-  }
-
-  // VisibleForTesting
-  @Nullable
-  InetAddress getLocalAddressForTest() {
-    return localAddress;
   }
 }
