+++ exporters/zipkin/src/test/java/io/opentelemetry/exporter/zipkin/ZipkinTestSpan.java
@@ -1,0 +1,87 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.exporter.zipkin;
+
+import static org.mockito.Mockito.mock;
+
+import io.opentelemetry.api.common.Attributes;
+import io.opentelemetry.api.trace.SpanContext;
+import io.opentelemetry.api.trace.SpanKind;
+import io.opentelemetry.api.trace.TraceFlags;
+import io.opentelemetry.api.trace.TraceState;
+import io.opentelemetry.sdk.resources.Resource;
+import io.opentelemetry.sdk.testing.trace.TestSpanData;
+import io.opentelemetry.sdk.trace.data.EventData;
+import io.opentelemetry.sdk.trace.data.StatusData;
+import io.opentelemetry.semconv.resource.attributes.ResourceAttributes;
+import java.net.InetAddress;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import zipkin2.Endpoint;
+import zipkin2.Span;
+
+public class ZipkinTestSpan {
+
+  static final String TRACE_ID = "d239036e7d5cec116b562147388b35bf";
+  static final String SPAN_ID = "9cc1e3049173be09";
+  static final String PARENT_SPAN_ID = "8b03ab423da481c5";
+
+  static final InetAddress localAddressForTesting = mock(InetAddress.class);
+
+  private static final Attributes attributes = Attributes.empty();
+  private static final List<EventData> annotations =
+      Collections.unmodifiableList(
+          Arrays.asList(
+              EventData.create(1505855799_433901068L, "RECEIVED", Attributes.empty()),
+              EventData.create(1505855799_459486280L, "SENT", Attributes.empty())));
+
+  static TestSpanData.Builder buildStandardSpan() {
+    return TestSpanData.builder()
+        .setSpanContext(
+            SpanContext.create(TRACE_ID, SPAN_ID, TraceFlags.getSampled(), TraceState.getDefault()))
+        .setParentSpanContext(
+            SpanContext.create(
+                TRACE_ID, PARENT_SPAN_ID, TraceFlags.getDefault(), TraceState.getDefault()))
+        .setResource(
+            Resource.create(
+                Attributes.builder().put(ResourceAttributes.SERVICE_NAME, "tweetiebird").build()))
+        .setStatus(StatusData.ok())
+        .setKind(SpanKind.SERVER)
+        .setName("Recv.helloworld.Greeter.SayHello")
+        .setStartEpochNanos(1505855794_194009601L)
+        .setEndEpochNanos(1505855799_465726528L)
+        .setAttributes(attributes)
+        .setTotalAttributeCount(attributes.size())
+        .setTotalRecordedEvents(annotations.size())
+        .setEvents(annotations)
+        .setLinks(Collections.emptyList())
+        .setHasEnded(true);
+  }
+
+  static Span buildZipkinSpan(Span.Kind kind) {
+    return standardZipkinSpanBuilder(kind).build();
+  }
+
+  static Span.Builder standardZipkinSpanBuilder(Span.Kind kind) {
+    return Span.newBuilder()
+        .traceId(TRACE_ID)
+        .parentId(PARENT_SPAN_ID)
+        .id(SPAN_ID)
+        .kind(kind)
+        .name("Recv.helloworld.Greeter.SayHello")
+        .timestamp(1505855794000000L + 194009601L / 1000)
+        .duration((1505855799000000L + 465726528L / 1000) - (1505855794000000L + 194009601L / 1000))
+        .localEndpoint(
+            Endpoint.newBuilder()
+                //                .ip(exporter.getLocalAddressForTest())
+                .ip(localAddressForTesting)
+                .serviceName("tweetiebird")
+                .build())
+        .addAnnotation(1505855799000000L + 433901068L / 1000, "RECEIVED")
+        .addAnnotation(1505855799000000L + 459486280L / 1000, "SENT");
+  }
+}
